Prep for custom codegen [1/n]
For liveness analysis we need CFG for functions, and for that we need to
be able to map blocks to predecessors and sucessors. This is possible to
do with a the previous Uniq type, but having a more dense representation
(where blocks always start from 0 in a function) allows using more
compact and efficient types like cranelift's PrimaryMap and
SecondaryMap.

Also fixes a bug in codegen where we'd previously use some locals before
declaring.